+++ src/transformers/models/mobilevitv2/modeling_mobilevitv2.py
@@ -131,7 +131,7 @@ class MobileViTV2InvertedResidual(nn.Module):
         self, config: MobileViTV2Config, in_channels: int, out_channels: int, stride: int, dilation: int = 1
     ) -> None:
         super().__init__()
-        expanded_channels = make_divisible(int(round(in_channels * config.expand_ratio)), 8)
+        expanded_channels = make_divisible(round(in_channels * config.expand_ratio), 8)
 
         if stride not in [1, 2]:
             raise ValueError(f"Invalid stride {stride}.")
