+++ src/transformers/models/musicgen_melody/modeling_musicgen_melody.py
@@ -1701,7 +1701,7 @@ class MusicgenMelodyForConditionalGeneration(PreTrainedModel, GenerationMixin):
 
                 # pad or truncate to config.chroma_length
                 if audio_hidden_states.shape[1] < self.config.chroma_length:
-                    n_repeat = int(math.ceil(self.config.chroma_length / audio_hidden_states.shape[1]))
+                    n_repeat = math.ceil(self.config.chroma_length / audio_hidden_states.shape[1])
                     audio_hidden_states = audio_hidden_states.repeat(1, n_repeat, 1)
                 else:
                     logger.warning(
@@ -1941,7 +1941,7 @@ class MusicgenMelodyForConditionalGeneration(PreTrainedModel, GenerationMixin):
 
             # pad or truncate to config.chroma_length
             if audio_hidden_states.shape[1] < self.config.chroma_length:
-                n_repeat = int(math.ceil(self.config.chroma_length / audio_hidden_states.shape[1]))
+                n_repeat = math.ceil(self.config.chroma_length / audio_hidden_states.shape[1])
                 audio_hidden_states = audio_hidden_states.repeat(1, n_repeat, 1)
             audio_hidden_states = audio_hidden_states[:, : self.config.chroma_length]
