+++ src/transformers/utils/model_parallel_utils.py
@@ -49,7 +49,7 @@ def assert_device_map(device_map, num_blocks):
 def get_device_map(n_layers, devices):
     """Returns a dictionary of layers distributed evenly across all devices."""
     layers = list(range(n_layers))
-    n_blocks = int(ceil(n_layers / len(devices)))
+    n_blocks = ceil(n_layers / len(devices))
     layers_list = [layers[i : i + n_blocks] for i in range(0, n_layers, n_blocks)]
 
     return dict(zip(devices, layers_list))
